% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{data}
\name{plotdumpdir}
\alias{plotdumpdir}
\title{Directory where generated plot and data files are written}
\format{
An object of class \code{character} of length 1.
}
\usage{
plotdumpdir
}
\description{
Set on package load to \code{/tmp/plotfiles}. Used by the rApache
handlers \code{\link{plotter}} and \code{\link{plotterpro}} as the
destination directory for generated PDF, PNG, CSV and TXT files.
}
\keyword{datasets}
